move utils into utils.rs
